big code cleanup
( code style, indent, trailling whitespace, doc comments, doc comments indent)

+++ src/rsfml/audio/sound/mod.rs
@@ -18,15 +18,15 @@
 *
 * 2. Altered source versions must be plainly marked as such, and must not be
 *    misrepresented as being the original software.
-* 
+*
 * 3. This notice may not be removed or altered from any source distribution.
 */
 
 /*!
-* Play sounds.
-*
-* Regular sound that can be played in the audio environment.
-*/
+ * Play sounds.
+ *
+ * Regular sound that can be played in the audio environment.
+ */
 
 use std::libc::c_float;
 use std::cast;
@@ -42,10 +42,10 @@ use ffi = ffi::audio::sound;
 pub mod rc;
 
 /**
-* Play sounds.
-*
-* Regular sound that can be played in the audio environment.
-*/
+ * Play sounds.
+ *
+ * Regular sound that can be played in the audio environment.
+ */
 pub struct Sound<'s> {
     #[doc(hidden)]
     priv sound :    *ffi::sfSound,
@@ -55,17 +55,17 @@ pub struct Sound<'s> {
 
 impl<'s> Sound<'s> {
     /**
-    * Create a new Sound
-    *
-    * Return Some(Sound) or None
-    */
+     * Create a new Sound
+     *
+     * Return Some(Sound) or None
+     */
     pub fn new() -> Option<Sound<'s>> {
         let s = unsafe {ffi::sfSound_create()};
         if s.is_null() {
             None
         }
         else {
-            Some(Sound { 
+            Some(Sound {
                 sound :     s,
                 buffer :    None
             })
@@ -73,10 +73,10 @@ impl<'s> Sound<'s> {
     }
 
     /**
-    * Create a new Sound
-    *
-    * Return Some(Sound) or None
-    */
+     * Create a new Sound
+     *
+     * Return Some(Sound) or None
+     */
     pub fn new_with_buffer(buffer : &'s SoundBuffer) -> Option<Sound<'s>> {
         let s = unsafe {ffi::sfSound_create()};
         if s.is_null() {
@@ -94,10 +94,10 @@ impl<'s> Sound<'s> {
     }
 
     /**
-    * Create a new sound by copying an existing one
-    *
-    * Return Some(Sound) or None
-    */
+     * Create a new sound by copying an existing one
+     *
+     * Return Some(Sound) or None
+     */
     pub fn clone(&self) -> Option<Sound<'s>> {
         let s = unsafe {ffi::sfSound_copy(self.sound)};
         if s.is_null() {
@@ -113,10 +113,10 @@ impl<'s> Sound<'s> {
     }
 
     /**
-    * Tell whether or not a sound is in loop mode
-    *
-    * Return true if the sound is looping, false otherwise
-    */
+     * Tell whether or not a sound is in loop mode
+     *
+     * Return true if the sound is looping, false otherwise
+     */
     pub fn set_loop(&mut self, lloop : bool) -> () {
         unsafe {
             if lloop == true {
@@ -129,10 +129,10 @@ impl<'s> Sound<'s> {
     }
 
     /**
-    * Tell whether or not a sound is in loop mode
-    *
-    * Return true if the sound is looping, false otherwise
-    */
+     * Tell whether or not a sound is in loop mode
+     *
+     * Return true if the sound is looping, false otherwise
+     */
     pub fn get_loop(&self) -> bool {
         match unsafe {ffi::sfSound_getLoop(self.sound)} {
             SFFALSE => false,
@@ -141,154 +141,153 @@ impl<'s> Sound<'s> {
     }
 
     /**
-    * Start or resume playing a sound
-    *
-    * This function starts the sound if it was stopped, resumes
-    * it if it was paused, and restarts it from beginning if it
-    * was it already playing.
-    * This function uses its own thread so that it doesn't block
-    * the rest of the program while the sound is played.
-    */
+     * Start or resume playing a sound
+     *
+     * This function starts the sound if it was stopped, resumes
+     * it if it was paused, and restarts it from beginning if it
+     * was it already playing.
+     * This function uses its own thread so that it doesn't block
+     * the rest of the program while the sound is played.
+     */
     pub fn play(&mut self) -> () {
         unsafe {ffi::sfSound_play(self.sound)}
     }
 
     /**
-    * Pause a sound
-    *
-    * This function pauses the sound if it was playing,
-    * otherwise (sound already paused or stopped) it has no effect.
-    */
+     * Pause a sound
+     *
+     * This function pauses the sound if it was playing,
+     * otherwise (sound already paused or stopped) it has no effect.
+     */
     pub fn pause(&mut self) -> () {
         unsafe {ffi::sfSound_pause(self.sound)}
     }
 
     /**
-    * Stop playing a sound
-    *
-    * This function stops the sound if it was playing or paused,
-    * and does nothing if it was already stopped.
-    * It also resets the playing position (unlike pause).
-    */
+     * Stop playing a sound
+     *
+     * This function stops the sound if it was playing or paused,
+     * and does nothing if it was already stopped.
+     * It also resets the playing position (unlike pause).
+     */
     pub fn stop(&mut self) -> () {
         unsafe {ffi::sfSound_stop(self.sound)}
     }
 
     /**
-    * Get the current status of a sound (stopped, paused, playing)
-    *
-    * Return current status
-    */
+     * Get the current status of a sound (stopped, paused, playing)
+     *
+     * Return current status
+     */
     pub fn get_status(&self) -> Status {
-        unsafe {cast::transmute(ffi::sfSound_getStatus(self.sound) as i8)}
+        unsafe { cast::transmute(ffi::sfSound_getStatus(self.sound)) }
     }
 
     /**
-    * Get the current playing position of a sound
-    *
-    * Return the current playing position
-    */
+     * Get the current playing position of a sound
+     *
+     * Return the current playing position
+     */
     pub fn get_playing_offset(&self) -> Time {
         Wrappable::wrap( unsafe {ffi::sfSound_getPlayingOffset(self.sound)})
     }
 
     /**
-    * Set the pitch of a sound
-    *
-    * The pitch represents the perceived fundamental frequency
-    * of a sound; thus you can make a sound more acute or grave
-    * by changing its pitch. A side effect of changing the pitch
-    * is to modify the playing speed of the sound as well.
-    * The default value for the pitch is 1.
-    *
-    * # Arguments
-    * * pitch - new pitch to apply to the sound
-    */
+     * Set the pitch of a sound
+     *
+     * The pitch represents the perceived fundamental frequency
+     * of a sound; thus you can make a sound more acute or grave
+     * by changing its pitch. A side effect of changing the pitch
+     * is to modify the playing speed of the sound as well.
+     * The default value for the pitch is 1.
+     *
+     * # Arguments
+     * * pitch - new pitch to apply to the sound
+     */
     pub fn set_pitch(&mut self, pitch : f32) -> () {
         unsafe {ffi::sfSound_setPitch(self.sound, pitch as c_float)}
     }
 
     /**
-    * Set the volume of a sound
-    *
-    * he volume is a value between 0 (mute) and 100 (full volume).
-    * The default value for the volume is 100.
-    *
-    * # Arguments
-    * * volume - Volume of the sound
-    */
+     * Set the volume of a sound
+     *
+     * he volume is a value between 0 (mute) and 100 (full volume).
+     * The default value for the volume is 100.
+     *
+     * # Arguments
+     * * volume - Volume of the sound
+     */
     pub fn set_volume(&mut self, volume : f32) -> () {
         unsafe {ffi::sfSound_setVolume(self.sound, volume as c_float)}
     }
 
     /**
-    * Make a sounds's position relative to the listener or absolute
-    *
-    * Making a sound relative to the listener will ensure that it will always
-    * be played the same way regardless the position of the listener.
-    * This can be useful for non-spatialized sounds, sounds that are
-    * produced by the listener, or sounds attached to it.
-    * The default value is false (position is absolute).
-    *
-    * # Arguments
-    * * relative - true to set the position relative, false to set it absolute
-    */
+     * Make a sounds's position relative to the listener or absolute
+     *
+     * Making a sound relative to the listener will ensure that it will always
+     * be played the same way regardless the position of the listener.
+     * This can be useful for non-spatialized sounds, sounds that are
+     * produced by the listener, or sounds attached to it.
+     * The default value is false (position is absolute).
+     *
+     * # Arguments
+     * * relative - true to set the position relative, false to set it absolute
+     */
     pub fn set_relative_to_listener(&mut self, relative : bool) -> () {
         unsafe {
             if relative == true {
                 ffi::sfSound_setRelativeToListener(self.sound, SFTRUE);
-            }
-            else {
+            } else {
                 ffi::sfSound_setRelativeToListener(self.sound, SFFALSE);
             }
         }
     }
 
     /**
-    * Set the minimum distance of a sound
-    *
-    * The "minimum distance" of a sound is the maximum
-    * distance at which it is heard at its maximum volume. Further
-    * than the minimum distance, it will start to fade out according
-    * to its attenuation factor. A value of 0 ("inside the head
-    * of the listener") is an invalid value and is forbidden.
-    * The default value of the minimum distance is 1.
-    *
-    * # Arguments
-    * * distance - New minimum distance of the sound
-    */
+     * Set the minimum distance of a sound
+     *
+     * The "minimum distance" of a sound is the maximum
+     * distance at which it is heard at its maximum volume. Further
+     * than the minimum distance, it will start to fade out according
+     * to its attenuation factor. A value of 0 ("inside the head
+     * of the listener") is an invalid value and is forbidden.
+     * The default value of the minimum distance is 1.
+     *
+     * # Arguments
+     * * distance - New minimum distance of the sound
+     */
     pub fn set_min_distance(&mut self, distance : f32) -> () {
         unsafe {ffi::sfSound_setMinDistance(self.sound, distance as c_float)}
     }
 
     /**
-    *  Set the attenuation factor of a sound
-    *
-    * The attenuation is a multiplicative factor which makes
-    * the sound more or less loud according to its distance
-    * from the listener. An attenuation of 0 will produce a
-    * non-attenuated sound, i.e. its volume will always be the same
-    * whether it is heard from near or from far. On the other hand,
-    * an attenuation value such as 100 will make the sound fade out
-    * very quickly as it gets further from the listener.
-    * The default value of the attenuation is 1.
-    *
-    * # Arguments
-    * * attenuation - New attenuation factor of the sound
-    */
+     *  Set the attenuation factor of a sound
+     *
+     * The attenuation is a multiplicative factor which makes
+     * the sound more or less loud according to its distance
+     * from the listener. An attenuation of 0 will produce a
+     * non-attenuated sound, i.e. its volume will always be the same
+     * whether it is heard from near or from far. On the other hand,
+     * an attenuation value such as 100 will make the sound fade out
+     * very quickly as it gets further from the listener.
+     * The default value of the attenuation is 1.
+     *
+     * # Arguments
+     * * attenuation - New attenuation factor of the sound
+     */
     pub fn set_attenuation(&mut self, attenuation : f32) -> () {
         unsafe {ffi::sfSound_setAttenuation(self.sound, attenuation as c_float)}
     }
 
     /**
-    * Change the current playing position of a sound
-    *
-    * The playing position can be changed when the sound is
-    * either paused or playing.
-    *
-    * # Arguments
-    * * timeOffset - New playing position
-    */
+     * Change the current playing position of a sound
+     *
+     * The playing position can be changed when the sound is
+     * either paused or playing.
+     *
+     * # Arguments
+     * * timeOffset - New playing position
+     */
     pub fn set_playing_offset(&mut self, timeOffset : Time) -> () {
         unsafe {
             ffi::sfSound_setPlayingOffset(self.sound, timeOffset.unwrap())
@@ -296,10 +295,10 @@ impl<'s> Sound<'s> {
     }
 
     /**
-    * Get the pitch of a sound
-    *
-    * Return the pitch of the sound
-    */
+     * Get the pitch of a sound
+     *
+     * Return the pitch of the sound
+     */
     pub fn get_pitch(&self) -> f32 {
         unsafe {
             ffi::sfSound_getPitch(self.sound) as f32
@@ -307,10 +306,10 @@ impl<'s> Sound<'s> {
     }
 
     /**
-    * Get the volume of a sound
-    *
-    * Return the volume of the sound, in the range [0, 100]
-    */
+     * Get the volume of a sound
+     *
+     * Return the volume of the sound, in the range [0, 100]
+     */
     pub fn get_volume(&self) -> f32 {
         unsafe {
             ffi::sfSound_getVolume(self.sound) as f32
@@ -318,10 +317,10 @@ impl<'s> Sound<'s> {
     }
 
     /**
-    * Tell whether a sound's position is relative to the listener or is absolute
-    *
-    * Return true if the position is relative, false if it's absolute
-    */
+     * Tell whether a sound's position is relative to the listener or is absolute
+     *
+     * Return true if the position is relative, false if it's absolute
+     */
     pub fn is_relative_to_listener(&self) -> bool {
         match unsafe {ffi::sfSound_isRelativeToListener(self.sound)} {
             SFFALSE => false,
@@ -330,10 +329,10 @@ impl<'s> Sound<'s> {
     }
 
     /**
-    * Get the minimum distance of a sound
-    *
-    * Return the minimum distance of the sound
-    */
+     * Get the minimum distance of a sound
+     *
+     * Return the minimum distance of the sound
+     */
     pub fn get_min_distance(&self) -> f32 {
         unsafe {
            ffi::sfSound_getMinDistance(self.sound) as f32
@@ -341,10 +340,10 @@ impl<'s> Sound<'s> {
     }
 
     /**
-    * Get the attenuation factor of a sound
-    *
-    * Return the attenuation factor of the sound
-    */
+     * Get the attenuation factor of a sound
+     *
+     * Return the attenuation factor of the sound
+     */
     pub fn get_attenuation(&self) -> f32 {
         unsafe {
             ffi::sfSound_getAttenuation(self.sound) as f32
@@ -352,15 +351,15 @@ impl<'s> Sound<'s> {
     }
 
     /**
-    * Set the source buffer containing the audio data to play
-    *
-    * It is important to note that the sound buffer is not copied,
-    * thus the sfSoundBuffer object must remain alive as long
-    * as it is attached to the sound.
-    *
-    * # Arguments
-    * * buffer - Sound buffer to attach to the sound
-    */
+     * Set the source buffer containing the audio data to play
+     *
+     * It is important to note that the sound buffer is not copied,
+     * thus the sfSoundBuffer object must remain alive as long
+     * as it is attached to the sound.
+     *
+     * # Arguments
+     * * buffer - Sound buffer to attach to the sound
+     */
     pub fn set_buffer(&mut self, buffer : &'s SoundBuffer) -> () {
         self.buffer = Some(buffer);
         unsafe {
@@ -369,19 +368,19 @@ impl<'s> Sound<'s> {
     }
 
     /**
-    * Get the audio buffer attached to a sound
-    *
-    * Return an option to Sound buffer attached to the sound or None
-    */
+     * Get the audio buffer attached to a sound
+     *
+     * Return an option to Sound buffer attached to the sound or None
+     */
     pub fn get_buffer(&self) -> Option<&'s SoundBuffer> {
         self.buffer
     }
 
     /**
-    * Get the 3D position of a sound in the audio scene
-    *
-    * Return the position of the sound in the world
-    */
+     * Get the 3D position of a sound in the audio scene
+     *
+     * Return the position of the sound in the world
+     */
     pub fn get_position(&self) -> Vector3f {
         unsafe {
             ffi::sfSound_getPosition(self.sound)
@@ -389,15 +388,15 @@ impl<'s> Sound<'s> {
     }
 
     /**
-    * Set the 3D position of a sound in the audio scene
-    *
-    * Only sounds with one channel (mono sounds) can be
-    * spatialized.
-    * The default position of a sound is (0, 0, 0).
-    *
-    * # Arguments
-    * * position - Position of the sound in the scene
-    */
+     * Set the 3D position of a sound in the audio scene
+     *
+     * Only sounds with one channel (mono sounds) can be
+     * spatialized.
+     * The default position of a sound is (0, 0, 0).
+     *
+     * # Arguments
+     * * position - Position of the sound in the scene
+     */
     pub fn set_position(&mut self, position : &Vector3f) -> () {
         unsafe {
             ffi::sfSound_setPosition(self.sound, *position)
@@ -405,17 +404,17 @@ impl<'s> Sound<'s> {
     }
 
     /**
-    * Set the 3D position of a sound in the audio scene
-    *
-    * Only sounds with one channel (mono sounds) can be
-    * spatialized.
-    * The default position of a sound is (0, 0, 0).
-    *
-    * # Arguments
-    * * x - X coordinate of the position of the sound in the scene
-    * * y - Y coordinate of the position of the sound in the scene
-    * * z - Z coordinate of the position of the sound in the scene
-    */
+     * Set the 3D position of a sound in the audio scene
+     *
+     * Only sounds with one channel (mono sounds) can be
+     * spatialized.
+     * The default position of a sound is (0, 0, 0).
+     *
+     * # Arguments
+     * * x - X coordinate of the position of the sound in the scene
+     * * y - Y coordinate of the position of the sound in the scene
+     * * z - Z coordinate of the position of the sound in the scene
+     */
     pub fn set_position3f(&mut self, x : f32, y : f32, z: f32) -> () {
         unsafe {
             ffi::sfSound_setPosition(self.sound, Vector3f::new(x, y, z))
